test finish

+++ japsswa/src/main/java/com/mes/jss/material/domain/MrOrderVO.java
@@ -1,8 +1,6 @@
 package com.mes.jss.material.domain;
 
-import java.util.Date;
-
-import org.springframework.format.annotation.DateTimeFormat;
+import java.math.BigDecimal;
 
 import lombok.Data;
 
@@ -13,8 +11,8 @@ public class MrOrderVO {	// 자재 발주
 	 String customerName;
 	 long empNo;
 	 String empName;
-	 long allCount;
-	 long productCount;
+	 BigDecimal allCount;
+	 BigDecimal productCount;
 	
 	 String mrOrderDate;
 }
